+++ examples/05-interoperability/07-converting-blocks-to-odt/src/App.tsx
@@ -92,6 +92,7 @@ export default function App() {
           backgroundColor: "red",
         },
       },
+      { type: "divider" },
       {
         type: "paragraph",
         content: [