+++ tests/src/unit/core/formatConversion/parse/parseTestInstances.ts
@@ -816,6 +816,15 @@ With Hard Break</p>
   },
   {
     testCase: {
+      name: "codeBlocksMultiLine",
+      content: `<pre><code>console.log("First Line")
+console.log("Second Line")
+console.log("Third Line")</code></pre>`,
+    },
+    executeTest: testParseHTML,
+  },
+  {
+    testCase: {
       name: "basicBlockquote",
       content: `<blockquote>This is a blockquote</blockquote>
 <p>This is not a blockquote</p>`,
@@ -922,6 +931,10 @@ With Hard Break</p>
     },
     executeTest: testParseHTML,
   },
+  {
+    testCase: { name: "divider", content: `<hr/>` },
+    executeTest: testParseHTML,
+  },
 ];
 
 export const parseTestInstancesMarkdown: TestInstance<
@@ -982,9 +995,15 @@ Paragraph
 
 Paragraph
 
+---
+
 P**ara***grap*h
 
+***
+
 P*ara*~~grap~~h
+
+___
 
 *   Bullet List Item
 
@@ -1075,6 +1094,20 @@ Regular paragraph`,
       content: `> **Bold** *Italic* ~~Strikethrough~~ ***Multiple***
 
 Regular paragraph`,
+    },
+    executeTest: testParseMarkdown,
+  },
+  {
+    testCase: {
+      name: "image",
+      content: `![Image](https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png)`,
+    },
+    executeTest: testParseMarkdown,
+  },
+  {
+    testCase: {
+      name: "video",
+      content: `![Video](https://interactive-examples.mdn.mozilla.net/media/cc0-videos/flower.webm)`,
     },
     executeTest: testParseMarkdown,
   },