+++ tests/src/unit/core/testSchema.ts
@@ -1,15 +1,14 @@
 import {
   BlockNoteSchema,
-  addNodeAndExtensionsToSpec,
+  createBlockConfig,
+  createBlockSpec,
   createImageBlockConfig,
   createImageBlockSpec,
   createInlineContentSpec,
   createPageBreakBlockSpec,
   createStyleSpec,
-  defaultBlockSpecs,
-  defaultInlineContentSpecs,
   defaultProps,
-  defaultStyleSpecs,
+  parseDefaultProps,
 } from "@blocknote/core";
 
 // BLOCKS ----------------------------------------------------------------------
@@ -17,12 +16,15 @@ import {
 // This is a modified version of the default image block that does not implement
 // a `toExternalHTML` function. It's used to test if the custom serializer by
 // default serializes custom blocks using their `render` function.
-const SimpleImage = addNodeAndExtensionsToSpec(
-  {
-    type: "simpleImage",
-    propSchema: createImageBlockConfig({}).propSchema,
-    content: "none",
-  },
+const SimpleImage = createBlockSpec(
+  createBlockConfig(
+    () =>
+      ({
+        type: "simpleImage",
+        propSchema: createImageBlockConfig({}).propSchema,
+        content: "none",
+      }) as const,
+  ),
   {
     render(block, editor) {
       return createImageBlockSpec().implementation.render.call(
@@ -34,13 +36,27 @@ const SimpleImage = addNodeAndExtensionsToSpec(
   },
 );
 
-const CustomParagraph = addNodeAndExtensionsToSpec(
+const CustomParagraph = createBlockSpec(
+  createBlockConfig(
+    () =>
+      ({
+        type: "customParagraph",
+        propSchema: defaultProps,
+        content: "inline",
+      }) as const,
+  ),
   {
-    type: "customParagraph",
-    propSchema: defaultProps,
-    content: "inline",
-  },
-  {
+    parse: (e) => {
+      if (e.tagName !== "P") {
+        return undefined;
+      }
+
+      if (e.classList.contains("custom-paragraph")) {
+        return parseDefaultProps(e);
+      }
+
+      return undefined;
+    },
     render: () => {
       const paragraph = document.createElement("p");
       paragraph.className = "custom-paragraph";
@@ -53,7 +69,6 @@ const CustomParagraph = addNodeAndExtensionsToSpec(
     toExternalHTML: () => {
       const paragraph = document.createElement("p");
       paragraph.className = "custom-paragraph";
-      paragraph.innerHTML = "Hello World";
 
       return {
         dom: paragraph,
@@ -62,12 +77,15 @@ const CustomParagraph = addNodeAndExtensionsToSpec(
   },
 );
 
-const SimpleCustomParagraph = addNodeAndExtensionsToSpec(
-  {
-    type: "simpleCustomParagraph",
-    propSchema: defaultProps,
-    content: "inline",
-  },
+const SimpleCustomParagraph = createBlockSpec(
+  createBlockConfig(
+    () =>
+      ({
+        type: "simpleCustomParagraph",
+        propSchema: defaultProps,
+        content: "inline",
+      }) as const,
+  ),
   {
     render: () => {
       const paragraph = document.createElement("p");
@@ -99,6 +117,7 @@ const Mention = createInlineContentSpec(
       dom.appendChild(document.createTextNode("@" + ic.props.user));
       dom.className = "mention-internal";
       dom.setAttribute("data-user", ic.props.user);
+      dom.style.backgroundColor = "red";
 
       return {
         dom,
@@ -113,6 +132,7 @@ const Mention = createInlineContentSpec(
       // Add attributes needed for round-trip compatibility
       dom.setAttribute("data-inline-content-type", "mention");
       dom.setAttribute("data-user", ic.props.user);
+      dom.style.backgroundColor = "red";
 
       return {
         dom,
@@ -196,21 +216,18 @@ const FontSize = createStyleSpec(
 
 // SCHEMA ----------------------------------------------------------------------
 
-export const testSchema = BlockNoteSchema.create({
+export const testSchema = BlockNoteSchema.create().extend({
   blockSpecs: {
-    ...defaultBlockSpecs,
     pageBreak: createPageBreakBlockSpec(),
-    customParagraph: CustomParagraph,
-    simpleCustomParagraph: SimpleCustomParagraph,
-    simpleImage: SimpleImage,
+    customParagraph: CustomParagraph(),
+    simpleCustomParagraph: SimpleCustomParagraph(),
+    simpleImage: SimpleImage(),
   },
   inlineContentSpecs: {
-    ...defaultInlineContentSpecs,
     mention: Mention,
     tag: Tag,
   },
   styleSpecs: {
-    ...defaultStyleSpecs,
     small: Small,
     fontSize: FontSize,
   },